policytree: Policy Learning via Doubly Robust Empirical Welfare Maximization over Trees

A package for learning simple rule-based policies, where the rule takes the form of a shallow decision tree. Applications include settings which require interpretable predictions, such as for example a medical treatment prescription. This package uses doubly robust reward estimates from grf to find a shallow, but globally optimal decision tree.

Some helpful links for getting started:

The R package documentation contains usage examples and method references (https://grf-labs.github.io/policytree/).

For community questions and answers around usage, see the GitHub issues page (https://github.com/grf-labs/policytree/issues).

See also

Examples

# \donttest{
# Multi-action policy learning example.
n <- 250
p <- 10
X <- matrix(rnorm(n * p), n, p)
W <- as.factor(sample(c("A", "B", "C"), n, replace = TRUE))
Y <- X[, 1] + X[, 2] * (W == "B") + X[, 3] * (W == "C") + runif(n)
multi.forest <- grf::multi_arm_causal_forest(X, Y, W)

# Compute doubly robust reward estimates.
Gamma.matrix <- double_robust_scores(multi.forest)

# Fit a depth 2 tree on a random training subset.
train <- sample(1:n, 200)
opt.tree <- policy_tree(X[train, ], Gamma.matrix[train, ], depth = 2)
opt.tree

#> policy_tree object 
#> Tree depth:  2 
#> Actions:  1: A 2: B 3: C 
#> Variable splits: 
#> (1) split_variable: X3  split_value: 0.114568 
#>   (2) split_variable: X2  split_value: 0.5315 
#>     (4) * action: 1 
#>     (5) * action: 2 
#>   (3) split_variable: X1  split_value: 1.4068 
#>     (6) * action: 3 
#>     (7) * action: 1 

# Predict treatment on held out data.
predict(opt.tree, X[-train, ])

#>  [1] 1 2 1 1 1 1 1 2 3 1 3 3 1 1 1 3 3 1 3 3 1 1 2 1 3 1 3 3 1 1 3 2 3 3 1 3 2 1
#> [39] 1 1 3 3 3 2 3 2 3 2 1 3

# }