Predict with multi_causal_forest

Computes estimates of \(\tau_a(x)\)

# S3 method for multi_causal_forest
predict(object, newdata = NULL, ...)

Arguments

object
The trained forest.

newdata
Points at which predictions should be made. If NULL, makes out-of-bag predictions on the training set instead (i.e., provides predictions at Xi using only trees that did not use the i-th training example). Note that this matrix should have the number of columns as the training matrix, and that the columns must appear in the same order.

...
Additional arguments passed to grf::predict.causal_forest.

Value

List containing matrix of predictions and other estimates (debiased error, etc.) for each treatment.

Examples

# \donttest{
# Train a multi causal forest.
n <- 250
p <- 10
X <- matrix(rnorm(n * p), n, p)
W <- sample(c("A", "B", "C"), n, replace = TRUE)
Y <- X[, 1] + X[, 2] * (W == "B") + X[, 3] * (W == "C") + runif(n)
multi.forest <- multi_causal_forest(X = X, Y = Y, W = W)

# Predict using the forest.
multi.forest.pred <- predict(multi.forest)
head(multi.forest.pred$predictions)

#>            A          B          C
#> 1  0.1744074 -0.4844184 -0.2399427
#> 2 -0.2101041 -0.1220175  0.8665341
#> 3 -0.2360677 -0.2330246  0.9307616
#> 4 -0.1455498  0.1174544  0.1860684
#> 5  0.2106911  0.3886140 -0.9179665
#> 6  0.1646506 -0.6161419 -0.1566698

# }