policytree: Policy Learning via Doubly Robust Empirical Welfare Maximization over Trees

A package for learning optimal policies via doubly robust empirical welfare maximization over trees. This package implements the multi-action doubly robust approach of Zhou et al. (2018) in the case where we want to learn policies that belong to the class of depth k decision trees. Many practical policy applications require interpretable predictions. For example, a drug prescription guide that follows a simple 2-question Yes/No checklist can be encoded as a depth 2 decision tree (does the patient have a heart condition - etc.). policytree currently has support for estimating multi-action treatment effects with one vs. all grf, calculating statistics such as double robust scores (support for a subset of grf forest types) and fitting optimal policies with exact tree search.

Some helpful links for getting started:

The R package documentation contains usage examples and method references.

For community questions and answers around usage, see the GitHub issues page.

See also

Author

Maintainer: Erik Sverdrup erikcs@stanford.edu

Authors:

Zhengyuan Zhou

Susan Athey

Stefan Wager

Ayush Kanodia

Examples

# \donttest{
# Multi-action treatment effect estimation
n <- 250
p <- 10
X <- matrix(rnorm(n * p), n, p)
W <- sample(c("A", "B", "C"), n, replace = TRUE)
Y <- X[, 1] + X[, 2] * (W == "B") + X[, 3] * (W == "C") + runif(n)
multi.forest <- multi_causal_forest(X = X, Y = Y, W = W)

# tau.hats
predict(multi.forest)$predictions

#>                 A            B            C
#> 1    0.1435295880  0.208221124 -0.906861730
#> 2   -0.0172739078  1.140138201 -1.151592263
#> 3   -0.1975272560 -0.650266915  0.872097893
#> 4   -0.0997803511  0.088638799  0.188553170
#> 5   -0.0539922634  1.226665180 -1.120508191
#> 6   -0.0555672511 -0.795877173  0.905655351
#> 7    0.2603439740 -0.482715894 -0.923031729
#> 8   -0.0520654173  0.456591669 -0.088462672
#> 9   -0.2108741852  0.350483487  0.489490761
#> 10   0.0397317980 -0.975033551  0.664657724
#> 11  -0.1204573289  1.114305077 -1.089109074
#> 12  -0.0442383949 -0.801341326  0.827793864
#> 13  -0.3056291966  0.825624832  0.303818017
#> 14   0.1065972293  0.022808268 -0.198984656
#> 15   0.1760312139 -0.638221550  0.181369613
#> 16  -0.1211035398 -0.133458352  0.363709187
#> 17  -0.1337373885 -0.342362195  0.748541586
#> 18  -0.1275578557  0.910175817 -0.036727441
#> 19  -0.1296566896 -0.121639142  0.549902440
#> 20   0.0542267642  1.022521293 -1.233438836
#> 21  -0.0228454727  1.216658289 -1.247971962
#> 22   0.0606613031  0.002185174 -0.115878774
#> 23   0.1482793399 -0.283035352 -0.010460036
#> 24   0.0602161063  0.252556238 -0.352490500
#> 25   0.1159301854 -0.932818937  0.584289712
#> 26  -0.1320293306 -0.196017168  0.451698823
#> 27  -0.1822714759  1.009542666 -0.187311510
#> 28  -0.2317008330  0.497782078  0.491270989
#> 29   0.2322126027 -0.594072753  0.013640546
#> 30   0.2368054240 -0.261684730 -0.739118888
#> 31   0.1072759214  0.214812597 -0.821283663
#> 32  -0.0178173899 -0.917017887  0.597194060
#> 33   0.1631958317  0.468350717 -1.199882646
#> 34  -0.0587232376  1.175838972 -1.205972521
#> 35  -0.2282873808  0.915734239  0.119281452
#> 36  -0.0069420437  0.938008201 -1.162416995
#> 37   0.0777443743 -0.804888718  0.366797301
#> 38   0.0929341331  0.359512269 -0.593833229
#> 39   0.2411767559  0.152514519 -1.092191701
#> 40   0.0665694337  0.300621402 -0.348400568
#> 41   0.2175330910 -0.569789036 -0.650280051
#> 42   0.0006981825 -0.107840377  0.257321097
#> 43   0.1769692152  0.032702811 -0.908599740
#> 44  -0.2326805487  0.888929856  0.148540072
#> 45  -0.1622087827 -0.174386154  0.563780558
#> 46   0.2021284973 -0.795197193  0.143833722
#> 47   0.1997362309 -0.800036064  0.181270402
#> 48  -0.1591617138  0.954849802 -0.036123037
#> 49  -0.0391881633 -0.763831579  0.680027322
#> 50  -0.0753048303 -0.368132785  0.518600369
#> 51   0.0403717158 -0.069419396  0.080100751
#> 52   0.2278179593 -0.366793481 -0.775384712
#> 53  -0.1183278310  0.522673122  0.097884608
#> 54   0.1162204182 -0.814214539  0.491406563
#> 55   0.2025153404 -0.555512678  0.105282315
#> 56  -0.1229557940 -0.174676741  0.457842423
#> 57  -0.0374599517 -0.782791098  0.833848394
#> 58  -0.1930676056  0.347197939  0.313277639
#> 59  -0.0370944371 -0.385597379  0.489364291
#> 60   0.2004325560 -0.789762100  0.187866422
#> 61   0.1939834553 -0.527354934  0.023557662
#> 62  -0.1193809554 -0.256703029  0.566999184
#> 63  -0.1054799545  1.169198988 -1.053637838
#> 64  -0.2627490686  1.013452581  0.297421759
#> 65  -0.3284662371  0.889622845  0.419484776
#> 66   0.1445717377 -0.277537195 -0.370099997
#> 67  -0.0821117789 -0.754783027  0.831140627
#> 68  -0.0878134369 -0.546895761  0.745445789
#> 69  -0.0659916568 -0.608790782  0.761200658
#> 70   0.1025222779  0.072143955 -0.254516742
#> 71  -0.0021161326 -1.020530960  0.873262067
#> 72  -0.0135277612 -0.230407175  0.383823466
#> 73   0.0737442767 -0.146577314 -0.066322185
#> 74   0.1231178277  0.552018603 -1.145650484
#> 75   0.1184134905  0.392525169 -1.117299869
#> 76   0.1124819041  0.361551352 -1.089929034
#> 77  -0.1457509138 -0.607764638  0.805727667
#> 78  -0.1191085235  0.039386169  0.255494898
#> 79   0.0812815082  0.020269593 -0.196605010
#> 80  -0.0061039887 -0.853677726  0.583191817
#> 81   0.1994630251 -0.487504943  0.079922818
#> 82   0.0439959844 -0.208999355  0.262957526
#> 83  -0.2647114135  0.867363811  0.325199415
#> 84  -0.0525198352 -0.567088323  0.712631402
#> 85   0.1585099954 -0.338732943 -0.620376480
#> 86  -0.1425310325 -0.209883621  0.506585271
#> 87   0.2045587900 -0.768794771  0.095694644
#> 88   0.0159411832 -0.050657658  0.078987654
#> 89  -0.0229721972 -0.993400165  0.899346655
#> 90  -0.0742847408  1.176429235 -1.196784683
#> 91  -0.1516427988  0.086675536  0.344911302
#> 92  -0.1788740947 -0.065010642  0.581464121
#> 93   0.0615837586 -0.026196256 -0.064254952
#> 94  -0.0790593380 -0.042427701  0.255758045
#> 95  -0.0505911633 -0.964645976  0.845508223
#> 96   0.1935156705 -0.019767479 -0.931969114
#> 97  -0.2963657872  0.898618015  0.356073435
#> 98  -0.2443035479  0.927036814  0.122269657
#> 99  -0.0639852582  0.817451247 -0.351782352
#> 100 -0.1043075271  0.027977669  0.237553245
#> 101  0.1326523382  0.334039343 -1.016957092
#> 102 -0.1491315091  1.025625449 -0.291139816
#> 103  0.1064387083  0.450351268 -1.029579842
#> 104 -0.0473409895 -0.853774755  0.848955530
#> 105  0.2251331289 -0.518908896 -0.743309704
#> 106 -0.1752600010  0.153087196  0.405780154
#> 107  0.0622681753 -0.738415815  0.543448130
#> 108 -0.0627463250  1.078571007 -0.716788397
#> 109  0.0177368848 -0.055347387  0.091944745
#> 110 -0.1616381825 -0.098503469  0.407968608
#> 111 -0.0850669769 -0.574960889  0.659905965
#> 112 -0.0068586931  1.039598136 -1.118531471
#> 113  0.0268443334 -0.190331777  0.181842349
#> 114  0.1187600705 -0.702423592  0.419360746
#> 115 -0.0301285548  0.905124907 -0.427085085
#> 116  0.1642845372 -0.072616066 -0.253753188
#> 117 -0.2808589236  0.939512437  0.321256216
#> 118 -0.0516967661 -0.930525310  0.809507746
#> 119 -0.1229901288 -0.736767922  0.844317889
#> 120  0.2408386426 -0.588877705 -0.607369001
#> 121 -0.0868108137 -0.366673158  0.691988126
#> 122  0.2494051126 -0.457738900 -0.750068988
#> 123 -0.0472788464  1.012412245 -0.341552936
#> 124 -0.0919911718  0.864298177 -0.391829095
#> 125 -0.0560984990  1.209830951 -1.123966979
#> 126 -0.2330099774  0.115877618  0.565747202
#> 127  0.0766556629  0.745476270 -1.220218157
#> 128 -0.0399384126 -0.986221572  0.699900007
#> 129  0.0895039823  0.153655251 -0.308099958
#> 130 -0.1381247357  0.850061742 -0.326384638
#> 131  0.1952224442 -0.549444088  0.187366700
#> 132  0.0190216935 -0.679715590  0.604869174
#> 133  0.1307773115  0.092164122 -0.308638849
#> 134 -0.0358457814  1.008847156 -1.175375090
#> 135 -0.0500157286 -0.978401159  0.754485056
#> 136 -0.0794009621  0.561553433 -0.158054079
#> 137  0.2103526692  0.281929824 -1.057649610
#> 138  0.2188282808 -0.593561646 -0.668481750
#> 139  0.1460123699  0.171092653 -1.020725353
#> 140  0.0984184035 -0.814475634  0.603125008
#> 141 -0.0506848037  1.156432799 -1.196685540
#> 142  0.1775053616 -0.369910227 -0.706453021
#> 143 -0.0817958939 -0.783396824  0.854749408
#> 144  0.0130303261  1.199865429 -1.239970073
#> 145 -0.0055335763  1.167070102 -1.075514269
#> 146 -0.0854813438 -0.783459679  0.891662973
#> 147 -0.0204440899  1.150878360 -1.016328549
#> 148  0.1918359115 -0.735351918  0.090989805
#> 149 -0.1780573592 -0.073059251  0.363507543
#> 150  0.2472914963 -0.353328779 -0.882036639
#> 151 -0.0728504832 -0.202048721  0.420630461
#> 152  0.1618773771 -0.142540882 -0.797778583
#> 153  0.0013154879  1.117226280 -1.205530427
#> 154  0.0282734877  0.601662561 -0.934522496
#> 155  0.0571679555  0.538175361 -1.123496139
#> 156  0.0245020202  0.104374642 -0.169905381
#> 157  0.2022127816 -0.272012153 -0.788207270
#> 158  0.1449182990 -0.918988621  0.469167724
#> 159 -0.2266404676  0.624472920  0.373655783
#> 160 -0.1060485616  0.925257385  0.027625119
#> 161  0.0827562689  0.247453382 -0.936018411
#> 162 -0.0616265261  1.225266249 -1.157133977
#> 163  0.2338049438 -0.304301804 -0.552197386
#> 164  0.1444071272  0.261296550 -0.931653813
#> 165 -0.0710302168 -0.723313767  0.680151860
#> 166  0.1159098045  0.390032381 -1.095247212
#> 167 -0.0173488083  0.594963049 -0.262203077
#> 168 -0.0193619233 -0.984386984  0.763005438
#> 169  0.0643553828 -0.777861803  0.566191365
#> 170 -0.1009035350 -0.602992157  0.823817651
#> 171 -0.0395709050  1.000801619 -1.200488297
#> 172  0.1495094708  0.091393544 -0.881199422
#> 173  0.0164033204  1.148521554 -1.106485354
#> 174 -0.0205466157 -1.002452750  0.682431296
#> 175 -0.1279409991  0.267165094  0.217513920
#> 176  0.1630054863 -0.766071620  0.208055063
#> 177 -0.1441292930  0.978743371 -0.107085073
#> 178 -0.1564546371  0.225198808  0.361694675
#> 179 -0.0301234812 -0.964846872  0.859330830
#> 180 -0.0788320840  0.964138468 -0.178279418
#> 181  0.1547533802  0.180212406 -0.955025187
#> 182  0.0134141423 -0.938097088  0.685544413
#> 183 -0.1466138383 -0.497813159  0.643134156
#> 184  0.0750661667  0.364500900 -0.664857331
#> 185  0.2729540631 -0.408515880 -0.834756025
#> 186 -0.0094790928  0.952382641 -0.584445609
#> 187 -0.1891560792  0.934576210 -0.091672360
#> 188 -0.2789793740  0.917442349  0.184754662
#> 189 -0.0406207051 -0.931485629  0.847430594
#> 190  0.2601446593 -0.366912988 -0.974388276
#> 191 -0.1376258024  0.684010640  0.003131385
#> 192 -0.1521770282 -0.168863559  0.617068440
#> 193  0.1503111507  0.248373198 -0.929363417
#> 194 -0.0873568736  1.145242165 -1.101278951
#> 195 -0.0891273823 -0.712918227  0.721165439
#> 196 -0.2667160991  0.546688525  0.399290715
#> 197  0.1071643047  0.223306142 -0.935105537
#> 198  0.0302671962 -0.909519985  0.581304365
#> 199  0.1990217466 -0.497637448 -0.503009968
#> 200  0.0957981632  0.321454472 -0.443307026
#> 201 -0.0014062149 -0.845024222  0.818363637
#> 202  0.1344591851  0.559330974 -1.231737645
#> 203  0.1623875515 -0.761280930  0.118166489
#> 204  0.0107416230 -0.219164999  0.308415815
#> 205 -0.2324829353  0.657981202  0.192666697
#> 206 -0.0196753757  1.250503744 -1.202528881
#> 207  0.0127394126 -0.022162455  0.151963494
#> 208 -0.0014658088  0.452037570 -0.157623988
#> 209  0.0638494259  0.762679423 -1.024376036
#> 210 -0.2436754233  0.499151544  0.410517152
#> 211  0.0148389797 -0.065068488  0.073089143
#> 212 -0.1149849869 -0.611351584  0.782729760
#> 213  0.1203822435  0.435577411 -1.022616795
#> 214  0.1746306421  0.333934407 -1.042340031
#> 215  0.0009646651  1.192985385 -1.203410555
#> 216  0.1313480886  0.132995398 -0.905914061
#> 217  0.2485523283 -0.744206066  0.112308618
#> 218 -0.1144357263 -0.684861683  0.654593231
#> 219 -0.1606748779  0.556311156  0.101461008
#> 220 -0.0745065735 -0.724341318  0.867782786
#> 221 -0.2893707046  0.840277641  0.481969880
#> 222 -0.0412884405 -0.871574080  0.860790640
#> 223 -0.2560325073  0.845346544  0.394488327
#> 224  0.1682057770 -0.670672226 -0.072101407
#> 225  0.1183584000 -0.938795461  0.609599662
#> 226 -0.0180504854  1.127162531 -1.216914223
#> 227 -0.2986153069  0.857128765  0.447934014
#> 228 -0.0138619847 -0.930371123  0.835404850
#> 229  0.2741169940 -0.412521683 -0.895299640
#> 230  0.0229067363  0.065464758 -0.037335073
#> 231  0.2112389661 -0.208450721 -0.780234296
#> 232  0.0765627983  0.853575168 -1.195316793
#> 233  0.0364691961  0.077988015 -0.072389708
#> 234 -0.0221849295 -0.871167280  0.663937018
#> 235  0.0782283559  0.057983277 -0.196911297
#> 236 -0.2036662751  0.732655970  0.068013937
#> 237  0.0639073385  0.387662082 -0.793925517
#> 238 -0.0828419918  0.919046194 -0.263369690
#> 239 -0.2902813823  0.871784777  0.431371603
#> 240 -0.1232579583 -0.649555303  0.812864472
#> 241 -0.0980349748  0.718203898 -0.169016707
#> 242 -0.0371500792 -0.972240136  0.740093604
#> 243 -0.2560642991  0.815035966  0.524905732
#> 244  0.1131851280  0.076624663 -0.269489092
#> 245 -0.0998088790 -0.844491236  0.843877725
#> 246 -0.1119790468 -0.655343779  0.643000925
#> 247  0.0477732237  0.075485928 -0.118505802
#> 248  0.0222124462  0.832350658 -0.896091660
#> 249 -0.0112522808  1.219631014 -1.237556232
#> 250 -0.2022397684  0.790583599  0.254245763

# Policy learning
Gamma.matrix <- double_robust_scores(multi.forest)

train <- sample(1:n, 200)
opt.tree <- policy_tree(X[train, ], Gamma.matrix[train, ], depth = 2)
opt.tree

#> policy_tree object 
#> Tree depth:  2 
#> Actions:  1: A 2: B 3: C 
#> Variable splits: 
#> (1) split_variable: X3  split_value: 0.21745 
#>   (2) split_variable: X2  split_value: -0.241997 
#>     (4) * action: 1 
#>     (5) * action: 2 
#>   (3) split_variable: X1  split_value: 1.61232 
#>     (6) * action: 3 
#>     (7) * action: 1 

# Predict treatment on held out data
predict(opt.tree, X[-train, ])

#>  [1] 2 3 1 3 1 3 2 1 1 2 2 3 3 2 1 3 2 1 3 3 2 2 3 2 3 1 3 2 2 3 3 3 3 3 1 2 1 2
#> [39] 2 2 2 1 3 1 3 2 2 1 3 2

# }